Move tests for use cases into separate folder tests/use_cases/
