+++ tests/use_cases/test_calculate_plan_expiration_and_check_if_expired.py
@@ -3,7 +3,8 @@ import datetime
 from arbeitszeit.use_cases import CalculatePlanExpirationAndCheckIfExpired
 from tests.data_generators import PlanGenerator
 from tests.datetime_service import FakeDatetimeService
-from tests.dependency_injection import injection_test
+
+from .dependency_injection import injection_test
 
 
 @injection_test
