+++ tests/use_cases/test_get_member_workplaces.py
@@ -1,7 +1,8 @@
 from arbeitszeit.use_cases import GetMemberWorkplaces
 from tests.data_generators import CompanyGenerator, MemberGenerator
-from tests.dependency_injection import injection_test
-from tests.repositories import CompanyWorkerRepository
+
+from .dependency_injection import injection_test
+from .repositories import CompanyWorkerRepository
 
 
 @injection_test
