+++ tests/use_cases/test_get_statistics.py
@@ -10,7 +10,8 @@ from tests.data_generators import (
     PlanGenerator,
 )
 from tests.datetime_service import FakeDatetimeService
-from tests.dependency_injection import injection_test
+
+from .dependency_injection import injection_test
 
 Number = Union[int, Decimal]
