+++ tests/use_cases/test_pay_consumer_product.py
@@ -2,9 +2,9 @@ import pytest
 
 from arbeitszeit import errors
 from arbeitszeit.use_cases import PayConsumerProduct
+from tests.data_generators import MemberGenerator, PlanGenerator
+from tests.datetime_service import FakeDatetimeService
 
-from .data_generators import MemberGenerator, PlanGenerator
-from .datetime_service import FakeDatetimeService
 from .dependency_injection import injection_test
 from .repositories import AccountRepository, TransactionRepository
