+++ tests/use_cases/test_pay_means_of_production.py
@@ -4,8 +4,9 @@ from arbeitszeit import errors
 from arbeitszeit.entities import PurposesOfPurchases
 from arbeitszeit.use_cases import PayMeansOfProduction
 from tests.data_generators import CompanyGenerator, PlanGenerator
-from tests.dependency_injection import injection_test
-from tests.repositories import AccountRepository, TransactionRepository
+
+from .dependency_injection import injection_test
+from .repositories import AccountRepository, TransactionRepository
 
 
 @injection_test
